made seprated file for upgraded components
