+++ org/Hibachi/client/src/form/form.module.ts
@@ -20,21 +20,24 @@ import {FileService} from "./services/fileservice";
 //form
 import {SWInput, SwInput} from "./components/swinput";
 import {SWFFormField} from "./components/swfformfield";
-import {SWForm,SwForm} from "./components/swform";
+import {SWForm} from "./components/swform";
+import {SwForm} from "./components/swform_upgraded";
 import {SWFForm} from "./components/swfform";
 import {SWFFileInput} from "./components/swffileinput";
-import {SWFormField, SwFormField} from "./components/swformfield";
+import {SWFormField} from "./components/swformfield";
+import {SwFormField} from "./components/swformfield_upgraded";
 import {SWFormFieldFile} from "./components/swformfieldfile";
 import {SWFormFieldJson} from "./components/swformfieldjson";
 import {SWFormFieldSearchSelect} from "./components/swformfieldsearchselect";
 import {SWFormRegistrar} from "./components/swformregistrar";
 import {SWErrorDisplay, SwErrorDisplay} from "./components/swerrordisplay";
 import {SWAddressForm} from "./components/swaddressform";
-import {SWPropertyDisplay, SwPropertyDisplay} from "./components/swpropertydisplay";
+import {SWPropertyDisplay} from "./components/swpropertydisplay";
+import {SwPropertyDisplay} from "./components/swpropertydisplay_upgraded";
 import {SWFPropertyDisplay} from "./components/swfpropertydisplay";
 import {SWFormSubscriber} from "./components/swformsubscriber";
 
-import {SwWorkflowBasic} from "../workflow/components/swworkflowbasic";
+import {SwWorkflowBasic} from "../workflow/components/swworkflowbasic_upgraded";
 
 @NgModule({
     declarations: [
